Add CPU/GPU support for int4 conversions in XLA.
int4 is currently only supported for convert instructions and a few instructions that do not produce any GPU code, such as tuple.

int4 is not yet supported in the Triton emitter, so using int4 will currently be slow on GPUs (and it is likely slow on CPUs as well).

int4 is packed on device but not on host. To represent int4 being packed, a pass, SubByteNormalization, sets element_size_in_bits() on every int4 layout to 4. Care is taken to ensure subsequent passes do not create int4 layouts with element_size_in_bits unset.

In LLVM IR, int4 is represented by i8, not i4, since arrays of i4 values are unpacked. When reading or writing the array using XLA's IrArray class, appropriate bitwise math is done to read the i8 value and mask out the correct bits.

PiperOrigin-RevId: 552665120