[tsl] forward TSL logging to Absl logging
Remove the LOG primitives in "default" and just rely on Absl.
These primitives were created when binary size was a concern.
That concern is not an issue anymore, so remove the primitives
and rely on Absl's logging.

Migrating users to depend on Absl directly is left as follow-up
work. New code must use Absl logging; this is enforced via the
linter's IWYU check.

PiperOrigin-RevId: 567642563