[XLA] [PjRT] Unify AOT mode and normal compilation for GPU
Current setup leads to excessive code duplication, as AOT relies on a separate
codepath entirely. I'd like to get AOT to a state where it seamlessly interops
with all other XLA tools, and reuses all the codepaths.

This CL instead of using a separate entry point, adds TargetConfig to
CompileOptions. If TargetConfig is present, the current GPU is not queried for
the compilation, and instead, the TargetConfig is used, emulating
cross-compiling for a given GPU.

Two recently added tests in PjRT were removed as they weren't correct:
StreamExecutorClient was serializing to a different proto than a parent class.

PiperOrigin-RevId: 571333724